Having to write another supyr_struct extension to fix UnionBlock
